move test to StringWriterTests.netcoreapp.cs
